Update SkiaSharp to latest.
Previously we specified SkiaSharp v2.88.5 because of a recent regression with JPEG output on non-windows.

But the older versions have a known vulnerability so let's go back to the using the latest version and just use PNG instead to avoid the bug where needed and add an explanatory comment.

Remove SkiaSharp from DisplayPDF sample which didn't actually use it.