Drop sample.
The sample was not designed for .NET, it was a carry over UI sample from the .NET Framework.

Now that we've completely dropped support for System.Drawing.Common this was no longer functional.

The DrawContents() overloads of our .NET library now only return SkiaSharp bitmap and it was causing it to look for the SkiaSharp reference which was just removed because no code used it.

The sample used Windows Forms which is only supported on Windows and not appropriate for cross platform development so this sample really didn't make much sense anymore.