Add nuget clean cache option.
Our new OCR api was added to DLE and the nuget test package was successfully put on the raid for these samples to use.

The problem is because the version is the same and it's already in our cache we don't update it, so let's add a new option to clean the cache for such a scenario.